OSIS-6272 : Move class into multiple files and Class-based views for OutsidePeriod and SelectTutor
